+++ DICOM/Imaging/DicomOverlayData.cs
@@ -228,6 +228,59 @@ namespace Dicom.Imaging {
 			if (ds.Contains(tag)) {
 				var elem = ds.FirstOrDefault(x => x.Tag == tag) as DicomElement;
 				_data = elem.Buffer;
+			} else {
+				// overlay embedded in high bits of pixel data
+				if (ds.InternalTransferSyntax.IsEncapsulated)
+					throw new DicomImagingException("Attempted to extract embedded overlay from compressed pixel data. Decompress pixel data before attempting this operation.");
+
+				var pixels = DicomPixelData.Create(ds);
+
+				// test that overlay fits in pixel data
+				if ((_originX + _columns) < pixels.Width && (_originY + _rows) < pixels.Height) {
+					var frame = pixels.GetFrame(0);
+
+					// calculate length of output buffer
+					var count = (_rows * _columns) / 8;
+					if (((_rows * _columns) % 8) != 0)
+					    count++;
+					if ((count & 1) != 0)
+						count++;
+
+					var bytes = new byte[count];
+					var bits = new BitArray(bytes);
+					int mask = 1 << _bitPosition;
+
+					if (pixels.BitsAllocated == 8) {
+						var data = ByteBufferEnumerator<byte>.Create(frame).ToArray();
+
+						int i = 0;
+						for (int y = _originY; y < (_originY + _rows); y++) {
+							for (int x = _originX; x < (_originX + _columns); x++) {
+								int n = (y * pixels.Width) + x;
+								if ((data[n] & mask) != 0)
+									bits[i] = true;
+								i++;
+							}
+						}
+					} else if (pixels.BitsAllocated == 16) {
+						// we don't really care if the pixel data is signed or not
+						var data = ByteBufferEnumerator<ushort>.Create(frame).ToArray();
+
+						int i = 0;
+						for (int y = _originY; y < (_originY + _rows); y++) {
+							for (int x = _originX; x < (_originX + _columns); x++) {
+								int n = (y * pixels.Width) + x;
+								if ((data[n] & mask) != 0)
+									bits[i] = true;
+								i++;
+							}
+						}
+					} else {
+						throw new DicomImagingException("Unable to extract embedded overlay from pixel data with bits stored greater than 16.");
+					}
+
+					_data = new MemoryByteBuffer(bytes);
+				}
 			}
 
 			_description = ds.Get<string>(OverlayTag(DicomTag.OverlayDescription), String.Empty);