+++ ChangeLog.md
@@ -6,6 +6,7 @@
 * Extension methods for recalculating and removing group length elements
 * Force calculation of group lengths when writing File Meta Info and Command datasets
 * Fix exception when attempting to display ROI overlays
+* Add ability to extract embedded overlays from pixel data
 
 #### v1.0.23 (09/26/2012)
 * Fix W/L calculation creating negative window width